ASC-127 Moved instrument class to core folder
