+++ atlast_sc/instruments/classes/Chai.py
@@ -1,6 +1,6 @@
 import re
 import astropy.units as u
-from atlast_sc.instrument import Instrument
+from atlast_sc.core.instrument import Instrument
 from atlast_sc.derived_groups import noise_temperature
 
 """
