+++ atlast_sc/instruments/classes/Default.py
@@ -1,6 +1,6 @@
 import astropy.units as u
 from astropy import constants
-from atlast_sc.instrument import Instrument
+from atlast_sc.core.instrument import Instrument
 from atlast_sc.derived_groups import noise_temperature
 
 """
