+++ atlast_sc/instruments/classes/Muscat.py
@@ -1,7 +1,7 @@
 from numpy import expm1, sqrt
 import astropy.units as u
 from astropy import constants
-from atlast_sc.instrument import Instrument
+from atlast_sc.core.instrument import Instrument
 from atlast_sc.derived_groups import noise_temperature
 
 """
