+++ atlast_sc/instruments/config.py
@@ -1,6 +1,6 @@
 import importlib, inspect, os
 from atlast_sc.core.utils import FileHelper
-from atlast_sc.instrument import Instrument
+from atlast_sc.core.instrument import Instrument
 
 class InstrumentConfig:
     """
